Implemented quicksearch for ids and code
